For SUPPORT-124 : [TABLE] After save data in a row the data isn't show in the list
The real problem was already fixedd by this morning's commits on case
SVY-10564.
This commit fixes an annoying visual effect that the selected record is
first scrolled to by showing it on bottom and then when selection is
increased it shows the next selected record but on top. So previous one
is no longer visible and can be confusing.