Implemented SVY-10650 : [servoy-extra table] Optimize rendering and loading more records when scrolling up
A few more fixes and improvements:
- fixed incorrect behavior in IE/Edge (the spacing divs on top and
bottom were not ok so scrolling/jumping to middle of large lists ...
didn't work well)
- sometimes you would end up with blank space (some rows not being
rendered when scrolling fast to bottom)
- sometimes it would take a long time to scroll to (unloaded/unrendered
yet) bottom if you had another page after that for large numbers of
rows, because it started to load rows from next page as well then it
tried to corrrect those and so on. Scroll know also jumped to various
places by 'correcting' incorrectly rendered rows due to this.