Related to SVY-19635: Foundsetpush issue
While making changes on the case, another existing bug in
servoy-extra-table surfaced (it could be reproduce even before the fixes
of this case if tab-panels were used to switch form instead of switching
main form).

If server-side would send a larger viewport (that did contain the
selection) initially to an extra-table with paging set to a fix > 0
value, then extra-table would correct it's viewport (load less records)
ignoring where the selected record was. It should have changed page
number as well in order to show selection initially as expected.

This bug surfaced because now even for main container form switches, the
client side component was created only after the data-changes - so it
didn't call setPreferredViewportSize() on server before form data was
sent to client...