we cant test for ng-reflect-xxx stuff
those are not there in none angular develop mode.
We always should test for the actual thing that happens in the DOM
but for value this is tricky because that is an internal tiny mce thing

maybe we can try to edit the value and then focus out and see if the
edit value will be emmitted.